dst/requirements.txt
aldjemy==0.8.0
awesome-slugify==1.6.5
barcode==0.6.0
beautifulsoup4==4.6.0
bs4==0.0.1
cairocffi==0.8.0
CairoSVG==2.1.3
certifi==2017.11.5
cffi==1.11.4
chardet==2.3.0
click==6.7
coreapi==2.3.3
coreapi-cli==1.0.8
coreschema==0.0.4
cssselect2==0.2.1
defusedxml==0.5.0
Django==2.0.1
django-appconf==1.0.2
django-ckeditor==5.4.0
django-cors-headers==2.1.0
django-debug-toolbar==1.9.1
django-extensions==1.9.8
django-imagekit==4.0.2
django-js-asset==0.1.1
django-pg-utils==0.1.5
django-ranged-response==0.2.0
django-simple-captcha==0.5.6
django-unused-media==0.1.10
djangorestframework==3.7.7
djangorestframework-jsonapi==2.3.1
fail2ban==0.9.6
future==0.16.0
html5lib==1.0.1
httplib2==0.9.2
inflection==0.3.1
itypes==1.1.0
Jinja2==2.10
MarkupSafe==1.0
mysqlclient==1.3.12
pdfrw==0.4
pilkit==2.0
Pillow==5.0.0
psycopg2==2.7.3.2
pyBarcode==0.7
pycparser==2.18
pycurl==7.43.0
pyinotify==0.9.6
PyMySQL==0.8.0
PyPDF2==1.26.0
Pyphen==0.9.4
python-apt==1.4.0b3
python-debian==0.1.30
python-debianbts==2.6.1
python-Levenshtein==0.12.0
python-telegram-bot==9.0.0
pytz==2017.3
qrcode==5.3
regex==2017.12.12
reportbug==7.1.7
requests==2.12.4
six==1.10.0
socketIO-client==0.7.2
SQLAlchemy==1.2.0
sqlparse==0.2.4
systemd-python==233
tinycss2==0.6.1
typing==3.6.2
Unidecode==0.4.21
uritemplate==3.0.0
urllib3==1.19.1
uWSGI==2.0.15
WeasyPrint==0.42
webencodings==0.5.1
websocket-client==0.46.0
xlrd==1.1.0
xlwt==1.3.0
xmltodict==0.11.0
